Fix size_t & do not ignore CFLAGS for test

+++ include/stddef.h
@@ -1,7 +1,7 @@
 #ifndef __LIBSEA_STDDEF_H__
 #define __LIBSEA_STDDEF_H__
 
-typedef unsigned long long size_t;
+typedef unsigned long size_t;
 
 #define NULL ((void*) 0)
